requirements.txt
yfinance
streamlit
scikit-learn
plotly-express